# Conflicts:
#	modules/aop/pom.xml
#	modules/apache-license-gen/pom.xml
#	modules/aws/pom.xml
#	modules/azure/pom.xml
#	modules/benchmarks/pom.xml
#	modules/cassandra/pom.xml
#	modules/cassandra/serializers/pom.xml
#	modules/cassandra/store/pom.xml
#	modules/clients/pom.xml
#	modules/cloud/pom.xml
#	modules/codegen/pom.xml
#	modules/compatibility/pom.xml
#	modules/compress/pom.xml
#	modules/control-utility/pom.xml
#	modules/core/pom.xml
#	modules/dev-utils/pom.xml
#	modules/direct-io/pom.xml
#	modules/extdata/platform/pom.xml
#	modules/gce/pom.xml
#	modules/geospatial/pom.xml
#	modules/hibernate-4.2/pom.xml
#	modules/hibernate-5.1/pom.xml
#	modules/hibernate-5.3/pom.xml
#	modules/hibernate-core/pom.xml
#	modules/ignored-tests/pom.xml
#	modules/indexing/pom.xml
#	modules/jcl/pom.xml
#	modules/jta/pom.xml
#	modules/kubernetes/pom.xml
#	modules/log4j/pom.xml
#	modules/log4j2/pom.xml
#	modules/mesos/pom.xml
#	modules/ml/catboost-model-parser/pom.xml
#	modules/ml/h2o-model-parser/pom.xml
#	modules/ml/pom.xml
#	modules/ml/spark-model-parser/pom.xml
#	modules/ml/xgboost-model-parser/pom.xml
#	modules/opencensus/pom.xml
#	modules/osgi/pom.xml
#	modules/rest-http/pom.xml
#	modules/scalar-2.10/pom.xml
#	modules/scalar/pom.xml
#	modules/schedule/pom.xml
#	modules/slf4j/pom.xml
#	modules/spark-2.4/pom.xml
#	modules/spark/pom.xml
#	modules/spring/pom.xml
#	modules/sqlline/pom.xml
#	modules/ssh/pom.xml
#	modules/tools/pom.xml
#	modules/urideploy/pom.xml
#	modules/visor-console-2.10/pom.xml
#	modules/visor-console/pom.xml
#	modules/visor-plugins/pom.xml
#	modules/web/ignite-appserver-test/pom.xml
#	modules/web/ignite-websphere-test/pom.xml
#	modules/web/pom.xml
#	modules/yardstick/pom.xml
#	modules/yarn/pom.xml
#	modules/zookeeper/pom.xml

+++ modules/core/src/main/java/org/apache/ignite/internal/util/GridJavaProcess.java
@@ -52,6 +52,9 @@ public final class GridJavaProcess {
     /** Internal protocol message prefix saying that the next text in the outputted line is pid. */
     public static final String PID_MSG_PREFIX = "my_pid_is:";
 
+    /** Default pid. */
+    private static final String DFLT_PID = "-1";
+
     /** Logger */
     private IgniteLogger log;
 
@@ -59,7 +62,7 @@ public final class GridJavaProcess {
     private Process proc;
 
     /** Pid of wrapped process. Made as array to be changeable in nested static class. */
-    private volatile String pid = "-1";
+    private volatile String pid = DFLT_PID;
 
     /** system.out stream grabber for process in which user class is running. */
     private ProcessStreamGrabber osGrabber;
@@ -186,17 +189,19 @@ public final class GridJavaProcess {
      * @throws Exception If any problem occurred.
      */
     public void kill() throws Exception {
-        Process killProc = U.isWindows() ?
-            Runtime.getRuntime().exec(new String[] {"taskkill", "/pid", pid, "/f", "/t"}) :
-            Runtime.getRuntime().exec(new String[] {"kill", "-9", pid});
+        if (!pid.equals(DFLT_PID)) {
+            Process killProc = U.isWindows() ?
+                    Runtime.getRuntime().exec(new String[]{"taskkill", "/pid", pid, "/f", "/t"}) :
+                    Runtime.getRuntime().exec(new String[]{"kill", "-9", pid});
 
-        if (!killProc.waitFor(5000, TimeUnit.MILLISECONDS))
-            throw new IllegalStateException("The kill process is hanging.");
+            if (!killProc.waitFor(5000, TimeUnit.MILLISECONDS))
+                throw new IllegalStateException("The kill process is hanging.");
 
-        int exitVal = killProc.exitValue();
+            int exitVal = killProc.exitValue();
 
-        if (exitVal != 0 && log.isInfoEnabled())
-            log.info(String.format("Abnormal exit value of %s for trying to kill the pid %s", exitVal, pid));
+            if (exitVal != 0 && log.isInfoEnabled())
+                log.info(String.format("Abnormal exit value of %s for trying to kill the pid %s", exitVal, pid));
+        }
 
         if (!proc.waitFor(5000, TimeUnit.MILLISECONDS))
             throw new IllegalStateException("Failed to kill grid java process.");
